% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{list_sc_heatmap_palettes}
\alias{list_sc_heatmap_palettes}
\title{List the built-in scHeatmap palettes}
\usage{
list_sc_heatmap_palettes()
}
\value{
A character vector of palette names accepted by
\code{\link[=sc_heatmap_palette]{sc_heatmap_palette()}}.
}
\description{
List the built-in scHeatmap palettes
}
